ft-highlight_enhance: fix bug of twice-selection of the whole lang tags when moving with keyboard arrow keys
The class attribute '_lastSelectedNode' holds the node value of the last selected mlang tag to compare with the current selection, to avoid selecting two times the tag when moving with arrow keys. This node value must be the one of the text, not the one of the span node, because the span selection (when the whole tag is selected) is triggered after the "simple" selection, so the comparation have to be done before.