issue #12: set submit listener to second optional submit button
The page may have more than one submit buttons, e.g., for saving and displaying, and for saving and returning to course. The easiest way to determine which event is the "triggerer", is to assign a different listener to each one, and this is the one for the second button that the page could have.