Make switching to/from HTML view work in IE
IE doesn't implement Function.caller.name (it's not part of
the standard), so we directly use the function from
the HTML view plugin object for the comparison.

Tested with Internet Explorer 11.0.9600.18762 in Windows 7 SP1,
and Firefox 56 and Chromium 61.0.3163.100 in Linux.

Also fix CSS bug introduced in a previous commit (missing ';'
at the end of the property).